Added my project 3 code to git hub
The encryption did not fully work but the concept of sending and receiving messages did.

I also did not fully understand the synchronized part of a function and that is probably not implemented correctly. (Ultimately my users were able to write data to the file without collision but that could be due to there only being 2 users (max 3) at any given time.